backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OCR & Image Processing
pytesseract==0.3.10
pillow==10.1.0
opencv-python-headless==4.8.1.78

# Utils
aiofiles==23.2.1
python-dateutil==2.8.2

# Development
httpx==0.25.2  # For testing